Moved reducers/actions into seperate files
