+++ src/Mailman.Server/ClientApp/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^3.4.0",
+    "@material-ui/core": "^3.9.2",
     "@material-ui/icons": "^3.0.1",
     "query-string": "^6.2.0",
     "react": "^16.6.1",
