+++ src/Mailman.Server/ClientApp/src/components/FetchData.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { actionCreators } from '../store/WeatherForecasts';
+import { actionCreators } from '../reducers/WeatherForecasts';
 
 class FetchData extends Component {
   componentWillMount() {
