+++ src/Mailman.Server/ClientApp/src/components/navigation/ResponsiveNavMenu.js
@@ -15,7 +15,7 @@ import { Link } from "react-router-dom";
 // State Management
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
-import { actionCreators } from "../../store/NavDrawer";
+import { actionCreators } from "../../reducers/NavDrawer";
 
 const drawerWidth = 240;
