+++ src/Mailman.Server/ClientApp/src/reducers/MergeTemplates.js
@@ -1,4 +1,9 @@
-import { REQUEST_MERGE_TEMPLATES, RECEIVE_MERGE_TEMPLATES } from './actionTypes';
+import {
+  REQUEST_MERGE_TEMPLATES,
+  RECEIVE_MERGE_TEMPLATES,
+  requestMergeTemplates,
+  receiveMergeTemplates
+} from '../actions/MergeTemplates';
 //import fetch from 'cross-fetch'; //most browsers don't natively support fetch yet, should install the npm package
 
 const initialState = {
@@ -42,44 +47,6 @@ const initialState = {
   sheetId: ''
 };
 
-
-const requestMergeTemplates = (spreadsheetId) => ( {
-  type: REQUEST_MERGE_TEMPLATES,
-  spreadsheetId
-});
-
-const receiveMergeTemplates = (spreadsheetId, json) => ({
-  type: RECEIVE_MERGE_TEMPLATES,
-  
-  payload: {
-    mergeTemplates: json,
-    spreadsheetId: spreadsheetId}
-
-});
-
-const config = {
-  method: 'GET',
-  headers: {
-    'Content-Type': 'application/json'
-  }
- };
-
-export function fetchMergeTemplates(spreadsheetId) {
-
-  return dispatch => {
-    dispatch(requestMergeTemplates(spreadsheetId));
-    return fetch(`https://localhost:5001/api/MergeTemplates/${spreadsheetId}`, config)
-      .then(response => {
-        console.log(response)
-        return response.json()})
-      .then(json => {
-        //debugger
-        console.log('Made it to this part! ', json[0]);
-        dispatch(receiveMergeTemplates(spreadsheetId, json))
-        return json;})
-  };
-}
-
 function handleErrors(response) {
   if (!response.ok) {
     console.log('Failed');
@@ -87,27 +54,6 @@ function handleErrors(response) {
   }
 
   return response;
-}
-
-function shouldFetchMergeTemplates() {
-  //TODO: make it so we don't issue a duplicate request 
-  //if isLoading is true, return false
-  //if we already have the data return false
-
-  return true;
-}
-
-export function fetchMergeTemplatesIfNeeded(spreadsheetId) {
- 
-  return (dispatch, getState) => {
-
-    if(shouldFetchMergeTemplates(spreadsheetId)) {
-      
-      return dispatch(fetchMergeTemplates(spreadsheetId))
-
-    }
-  }
-
 }
 
 export const reducer = (state, action) => { // state = initialState
