+++ src/Mailman.Server/ClientApp/src/reducers/NavDrawer.js
