+++ src/Mailman.Server/ClientApp/src/reducers/WeatherForecasts.js
