--- src/Mailman.Server/ClientApp/src/store/actionTypes.js
@@ -1,9 +1,0 @@
- export const REQUEST_MERGE_TEMPLATES = "REQUEST_MERGE_TEMPLATES";
- export const RECEIVE_MERGE_TEMPLATES = "RECEIVE_MERGE_TEMPLATES";
-//delete
-//start new
-
- //export const FETCH_MERGE_TEMPLATES_REQUEST   = 'FETCH_MERGE_TEMPLATES_REQUEST';
- //export const FETCH_MERGE_TEMPLATES_SUCCESS = 'FETCH_MERGE_TEMPLATES_SUCCESS';
- //export const FETCH_MERGE_TEMPLATES_FAILURE = 'FETCH_MERGE_TEMPLATES_FAILURE';
-
