+++ src/Mailman.Server/ClientApp/src/store/configureStore.js
@@ -1,10 +1,10 @@
 ﻿import { applyMiddleware, combineReducers, compose, createStore } from "redux";
 import thunk from "redux-thunk";
 import { routerReducer, routerMiddleware } from "react-router-redux";
-import * as Counter from "./Counter";
-import * as WeatherForecasts from "./WeatherForecasts";
-import * as NavDrawer from "./NavDrawer";
-import * as MergeTemplates from "./MergeTemplates";
+import * as Counter from "../reducers/Counter";
+import * as WeatherForecasts from "../reducers/WeatherForecasts";
+import * as NavDrawer from "../reducers/NavDrawer";
+import * as MergeTemplates from "../reducers/MergeTemplates";
 
 export default function configureStore(history, initialState) {
